Feat: Separate token kind from token literal, such that the kind can be matched irrespective of the literal.
This is useful when, for example, peeking the next token. One wants to
assert that the token is of the correct kind, but not care about the
literal associated with it.

Since when parsing identifiers and numbers one cannot know ahead of time
the literal value contained, it is impossible to match against, unless
kind and literal are separate values.